Adds 3D transformation and animation to dice
The dice-twenty element now displays 4 of the 20 sides of the 20-sided
dice. It uses 3D CSS transforms to approximate a 3D dice.

When you click roll, it cycles through a few random values to give the
impression of animating. This makes it clearer what has happened when
there are repeat values.